+++ src/main/java/com/yapp/artie/domain/user/application/port/in/GetUserThumbnailQuery.java
@@ -1,8 +1,6 @@
 package com.yapp.artie.domain.user.application.port.in;
 
-import com.yapp.artie.domain.user.dto.response.UserThumbnailResponseDto;
-
 public interface GetUserThumbnailQuery {
 
-  UserThumbnailResponseDto loadUserThumbnailById(Long id);
+  GetUserThumbnailResponse loadUserThumbnailById(Long id);
 }
