+++ src/main/java/com/yapp/artie/domain/user/application/port/in/RegisterUserResponse.java
@@ -1,16 +1,16 @@
-package com.yapp.artie.domain.user.dto.response;
+package com.yapp.artie.domain.user.application.port.in;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
 
 @Schema(description = "회원 Response")
 @Getter
-public class CreateUserResponseDto {
+public class RegisterUserResponse {
 
   @Schema(description = "아이디")
   public Long id;
 
-  public CreateUserResponseDto(Long id) {
+  public RegisterUserResponse(Long id) {
     this.id = id;
   }
 }
