+++ src/main/java/com/yapp/artie/domain/user/application/port/in/RegisterUserUseCase.java
@@ -1,8 +1,6 @@
 package com.yapp.artie.domain.user.application.port.in;
 
-import com.yapp.artie.domain.user.dto.response.CreateUserResponseDto;
-
 public interface RegisterUserUseCase {
 
-  CreateUserResponseDto register(String uid, String username, String picture);
+  RegisterUserResponse register(String uid, String username, String picture);
 }
